+++ app/models/transaction.py
@@ -13,7 +13,7 @@ class Transaction(db.Model):
     portfolio_id = db.Column(db.Integer, db.ForeignKey(
         add_prefix_for_prod("portfolios.id")), nullable=False)
     total_cost = db.Column(db.Float, nullable=False)
-    shares = db.Column(db.Integer, nullable=False)
+    shares = db.Column(db.Float, nullable=False)
     date = db.Column(db.DateTime, nullable=False)
 
     # relationships
